+++ fltk-sys/src/browser.rs
@@ -185,7 +185,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Widget_set_trigger(arg1: *mut Fl_Widget, arg2: libc::c_int);
+    pub fn Fl_Widget_set_when(arg1: *mut Fl_Widget, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Widget_when(arg1: *const Fl_Widget) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Widget_image(arg1: *const Fl_Widget) -> *mut libc::c_void;
@@ -450,7 +453,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Browser_set_trigger(arg1: *mut Fl_Browser, arg2: libc::c_int);
+    pub fn Fl_Browser_set_when(arg1: *mut Fl_Browser, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Browser_when(arg1: *const Fl_Browser) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Browser_image(arg1: *const Fl_Browser) -> *mut libc::c_void;
@@ -853,7 +859,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Hold_Browser_set_trigger(arg1: *mut Fl_Hold_Browser, arg2: libc::c_int);
+    pub fn Fl_Hold_Browser_set_when(arg1: *mut Fl_Hold_Browser, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Hold_Browser_when(arg1: *const Fl_Hold_Browser) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Hold_Browser_image(arg1: *const Fl_Hold_Browser) -> *mut libc::c_void;
@@ -1273,7 +1282,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Select_Browser_set_trigger(arg1: *mut Fl_Select_Browser, arg2: libc::c_int);
+    pub fn Fl_Select_Browser_set_when(arg1: *mut Fl_Select_Browser, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Select_Browser_when(arg1: *const Fl_Select_Browser) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Select_Browser_image(arg1: *const Fl_Select_Browser) -> *mut libc::c_void;
@@ -1705,7 +1717,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_Multi_Browser_set_trigger(arg1: *mut Fl_Multi_Browser, arg2: libc::c_int);
+    pub fn Fl_Multi_Browser_set_when(arg1: *mut Fl_Multi_Browser, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_Multi_Browser_when(arg1: *const Fl_Multi_Browser) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_Multi_Browser_image(arg1: *const Fl_Multi_Browser) -> *mut libc::c_void;
@@ -2130,7 +2145,10 @@ extern "C" {
     );
 }
 extern "C" {
-    pub fn Fl_File_Browser_set_trigger(arg1: *mut Fl_File_Browser, arg2: libc::c_int);
+    pub fn Fl_File_Browser_set_when(arg1: *mut Fl_File_Browser, arg2: libc::c_int);
+}
+extern "C" {
+    pub fn Fl_File_Browser_when(arg1: *const Fl_File_Browser) -> libc::c_int;
 }
 extern "C" {
     pub fn Fl_File_Browser_image(arg1: *const Fl_File_Browser) -> *mut libc::c_void;